It basically checks the randomly take digit within the range is even/odd, positive/negative and prime/composite number.
This  is a project where we run if else, statements to print if the randomly taken number from the range is positive or negative , similarly for even or odd. We combine for loop and if statement to generate a result if the random taken number is a  prime number or a composite number.